+++ sub-proj/JSToolJS/realjsformatter.js
@@ -118,8 +118,6 @@ class RealJSFormatter extends JSParser.JSParser {
 
         this.m_lineWaitVec = [];
 
-        this.m_blockStack = [];
-
         // stack used for solving loop in if
         this.m_brcNeedStack = []; // () after if
 
@@ -202,6 +200,7 @@ class RealJSFormatter extends JSParser.JSParser {
         let tokenBFirst = '';
 
         this.StartParse();
+
         while (this.GetToken()) {
             bHaveNewLine = false; // bHaveNewLine means there will be newline, m_bNewLine means already has newline
             tokenAFirst = this.m_tokenA.code.charAt(0);
@@ -785,12 +784,27 @@ class RealJSFormatter extends JSParser.JSParser {
             this.m_bAssign = true;
         }
 
+        if (this.m_tokenA.code.endsWith("${")) {
+            this.m_blockStack.push(JSParser.JS_TEMP_LITE);
+        }
+
+        if (JSParser.StackTopEq(this.m_blockStack, JSParser.JS_TEMP_LITE) && this.m_tokenA.code.startsWith("}")) {
+            this.m_blockStack.pop();
+        }
+
         if (this.m_tokenB.type == JSParser.STRING_TYPE ||
             this.m_tokenB.type == JSParser.COMMENT_TYPE_1 ||
             this.m_tokenB.type == JSParser.COMMENT_TYPE_2 ||
             this.m_tokenB.code == "{" ||
             (this.m_declareKeywordSet.includes(this.m_tokenA.code) && this.m_tokenB.code == "[")) {
-            this.PutToken(this.m_tokenA, "", " ");
+            if (this.m_tokenA.type == JSParser.STRING_TYPE &&
+                this.m_tokenA.code.endsWith("${")) {
+                this.m_bTemplatePut = true;
+                this.PutToken(this.m_tokenA);
+                this.m_bTemplatePut = false;
+            } else {
+                this.PutToken(this.m_tokenA, "", " ");
+            }
 
             //if(m_blockStack.top() != 't' && IsType(m_tokenA))
             //m_blockStack.push('t'); // variable statment
@@ -801,7 +815,8 @@ class RealJSFormatter extends JSParser.JSParser {
             (this.m_bracketKeywordSet.includes(this.m_tokenA.code) &&
                 this.m_tokenB.code != ";")) {
             this.PutToken(this.m_tokenA, "", " ");
-        } else if (this.m_tokenA.code[0] == '`' && this.m_tokenA.code[this.m_tokenA.code.length - 1] == '`') {
+        } else if ((this.m_tokenA.code.startsWith("`") || this.m_tokenA.code.startsWith("}")) &&
+            (this.m_tokenA.code.endsWith("`") || this.m_tokenA.code.endsWith("${"))) {
             this.m_bTemplatePut = true;
             this.PutToken(this.m_tokenA);
             this.m_bTemplatePut = false;