+++ sub-proj/Test/jsformattest.py
@@ -3,7 +3,6 @@
 # Author: Sun Junwen
 # Date: 2014-01-04
 #
-import hashlib
 import os
 import sys
 from subprocess import call
@@ -12,28 +11,17 @@ import comm_util
 from testbase import *
 
 TEST_CASE_DIR = 'jsformat'
+WIN_ARM64 = 'ARM64'
 
 JSFORMATTER_PATH_WIN = '../../trunk/debug/JSFormatterTest.exe'
 JSFORMATTER_REL_PATH_WIN = '../../trunk/release/JSFormatterTest.exe'
 JSFORMATTER_PATH_WIN_64 = '../../trunk/x64/debug/JSFormatterTest.exe'
 JSFORMATTER_REL_PATH_WIN_64 = '../../trunk/x64/release/JSFormatterTest.exe'
-JSFORMATTER_LIB_PATH_MAC = '../../trunk/DerivedData/JSTool/Build/Products/Debug'
-JSFORMATTER_LIB_REL_PATH_MAC = '../../trunk/DerivedData/JSTool/Build/Products/Release'
+JSFORMATTER_PATH_WIN_ARM64 = '../../trunk/arm64/debug/JSFormatterTest.exe'
+JSFORMATTER_REL_PATH_WIN_ARM64 = '../../trunk/arm64/release/JSFormatterTest.exe'
 JSFORMATTER_PATH_MAC = '../../trunk/DerivedData/JSTool/Build/Products/Debug/JSFormatterTest'
 JSFORMATTER_REL_PATH_MAC = '../../trunk/DerivedData/JSTool/Build/Products/Release/JSFormatterTest'
 JSFORMATTER_NODEJS_SCRIPT_PATH = '../JSToolJS/jsfjsnode.js'
-
-class MacOSCaseRuntime(CaseRuntime):
-    def __init__(self, runtime_path, lib_path):
-        super(MacOSCaseRuntime, self).__init__(runtime_path)
-        self.lib_path = lib_path
-        os.environ['DYLD_LIBRARY_PATH'] = self.lib_path
-
-    def dump_name(self):
-        comm_util.log_print('macOSCaseRuntime')
-
-    def dump_info(self):
-        comm_util.log_print('DYLD_LIBRARY_PATH=%s' % (os.environ['DYLD_LIBRARY_PATH']))
 
 class NodeCaseRuntime(CaseRuntime):
     def _case_execute(self, test_case):
@@ -88,6 +76,11 @@ def main():
     nodejs = False
     validate = False
 
+    win_arm64 = False
+    machine = comm_util.get_machine()
+    if comm_util.is_windows() and machine == WIN_ARM64:
+        win_arm64 = True
+
     for argv in sys.argv:
         argv = argv.lower()
         if argv == 'node' or argv == 'nodejs' or argv == 'js':
@@ -107,34 +100,31 @@ def main():
         if argv == '64' or argv == 'x64':
             x64 = True
         if argv == '32' or argv == 'x86':
-            if not comm_util.is_macos():
+            if not comm_util.is_macos() and not win_arm64:
                 x64 = False
 
     # system check
     if nodejs == False:
         if not comm_util.is_windows() and not comm_util.is_macos():
             if comm_util.is_linux():
-                comm_util.log_print('Only node support Linux.')
+                comm_util.log_print('Linux only support node.')
             else:
                 comm_util.log_print('Unknown operating system.')
             return
     else:
+        if win_arm64:
+            comm_util.log_print('Windows ARM64 not support Node.')
+            return
         if not comm_util.is_windows() and not comm_util.is_macos() and not comm_util.is_linux():
             comm_util.log_print('Unknown operating system.')
             return
 
     # prepare path
     jsformatter_path_sel = ''
-    jsformatter_lib_path_sel = ''
     jsformatter_nodejs_script_sel = ''
 
     if nodejs == False and validate == False:
-        if comm_util.is_macos():
-            jsformatter_lib_path_sel = JSFORMATTER_LIB_PATH_MAC
-            if release:
-                jsformatter_lib_path_sel = JSFORMATTER_LIB_REL_PATH_MAC
-
-        if comm_util.is_windows():
+        if comm_util.is_windows() and not win_arm64:
             jsformatter_path_sel = JSFORMATTER_PATH_WIN
             if release and x64:
                 jsformatter_path_sel = JSFORMATTER_REL_PATH_WIN_64
@@ -142,6 +132,10 @@ def main():
                 jsformatter_path_sel = JSFORMATTER_PATH_WIN_64
             elif release:
                 jsformatter_path_sel = JSFORMATTER_REL_PATH_WIN
+        if comm_util.is_windows() and win_arm64:
+            jsformatter_path_sel = JSFORMATTER_PATH_WIN_ARM64
+            if release:
+                jsformatter_path_sel = JSFORMATTER_REL_PATH_WIN_ARM64
         if comm_util.is_macos():
             jsformatter_path_sel = JSFORMATTER_PATH_MAC
             if release:
@@ -155,15 +149,12 @@ def main():
     # make runtime
     case_runtime = 0
     if nodejs == False and validate == False:
-        if comm_util.is_windows():
-            case_runtime = CaseRuntime(jsformatter_path_sel)
-        if comm_util.is_macos():
-            case_runtime = MacOSCaseRuntime(jsformatter_path_sel, jsformatter_lib_path_sel)
+        case_runtime = CaseRuntime(jsformatter_path_sel)
     if nodejs:
         case_runtime = NodeCaseRuntime(jsformatter_nodejs_script_sel)
     if validate:
-        if comm_util.is_macos():
-            comm_util.log_print('Validate only support Windows.')
+        if comm_util.is_macos() or win_arm64:
+            comm_util.log_print('Validate only support Windows x64.')
             return
         case_runtime = ValidateCaseRuntime(jsformatter_path_sel, jsformatter_nodejs_script_sel)